Refactored Daylio API & added list all entries route
